refactor: UserController 리팩토링 - 로그인 관련 컨트롤러 분리
회원가입, 로그인, 로그아웃, 회원탈퇴 컨트롤러 분리
회원탈퇴 매핑 변경: /user/quit
- securityconfig 추가, profile.html의 js 요청 url 수정
