+++ src/main/java/com/donguri/jejudorang/domain/user/api/UserController.java
@@ -1,9 +1,7 @@
 package com.donguri.jejudorang.domain.user.api;
 
-import com.donguri.jejudorang.domain.community.dto.response.CommunityListResponseDto;
-import com.donguri.jejudorang.domain.user.dto.request.*;
-import com.donguri.jejudorang.domain.user.dto.request.email.*;
-import com.donguri.jejudorang.domain.user.dto.response.ProfileResponse;
+import com.donguri.jejudorang.domain.user.dto.request.LoginRequest;
+import com.donguri.jejudorang.domain.user.dto.request.SignUpRequest;
 import com.donguri.jejudorang.domain.user.service.UserService;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
@@ -11,10 +9,6 @@ import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
@@ -22,79 +16,32 @@ import org.springframework.stereotype.Controller;
 import org.springframework.transaction.UnexpectedRollbackException;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CookieValue;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.util.Map;
 import java.util.Optional;
-
 
 @Slf4j
 @Controller
 @RequestMapping("/user")
 public class UserController {
 
-    @Value("${jwt.cookie-expire}")
-    private int cookieTime;
-
+    private final int cookieTime;
     @Autowired private final UserService userService;
-    public UserController(UserService userService) {
+
+    public UserController(@Value("${jwt.cookie-expire}") int cookieTime, UserService userService) {
+        this.cookieTime = cookieTime;
         this.userService = userService;
     }
 
-    /*
-    * 이메일 인증 번호 전송 (+ 중복 확인)
-    * */
-    @PostMapping("/email/verify")
-    public ResponseEntity<?> checkDuplicatedAndSendEmailCode(@RequestBody @Valid MailSendRequest mailSendRequest, BindingResult bindingResult) {
-        try {
-            checkValidationAndReturnException(bindingResult);
 
-            userService.checkMailDuplicatedAndSendVerifyCode(mailSendRequest);
-
-            log.info("이메일 인증 번호 전송 완료");
-            return new ResponseEntity<>(HttpStatus.OK);
-
-        } catch (NullPointerException e) {
-            log.error("이메일 인증 번호 전송 실패: {}", e.getMessage());
-            return new ResponseEntity<>("이메일을 입력해주세요.", HttpStatus.BAD_REQUEST);
-
-        } catch (Exception e) {
-            log.error("이메일 인증 번호 전송 실패: {}", e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-        }
-    }
 
     /*
-    * 이메일 인증 번호 확인
-    * */
-    @PostMapping("/email/verify-check")
-    public ResponseEntity<?> checkEmailCode(@RequestBody @Valid MailVerifyRequest mailVerifyRequest, BindingResult bindingResult) {
-        try {
-            checkValidationAndReturnException(bindingResult);
-
-            boolean checkRes = userService.checkVerifyMail(mailVerifyRequest);
-            if (checkRes) {
-                log.info("이메일 인증 완료 : 인증 번호 일치");
-                return new ResponseEntity<>(HttpStatus.OK);
-
-            } else {
-                log.error("이메일 인증 실패 : 인증 번호 불일치");
-                return new ResponseEntity<>("인증 번호가 일치하지 않습니다", HttpStatus.BAD_REQUEST);
-            }
-
-        } catch (NullPointerException e) {
-            log.error("이메일 인증 실패: {}", e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-
-        } catch (Exception e) {
-            log.error(e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_GATEWAY);
-        }
-    }
-
-    /*
-    * 회원 가입
-    * */
+     * 회원 가입
+     * */
     @GetMapping("/signup")
     public String registerForm() {
         return "/user/login/signUpForm";
@@ -115,10 +62,17 @@ public class UserController {
         }
     }
 
+
+    /*
+    * 로그인
+    * GET, POST
+    *
+    * */
     @GetMapping("/login")
     public String signInForm() {
         return "/user/login/signInForm";
     }
+
     @PostMapping("/login")
     public String authenticateUser(@Valid LoginRequest loginRequest, BindingResult bindingResult,
                                    HttpServletResponse response, Model model) {
@@ -152,20 +106,12 @@ public class UserController {
         }
     }
 
-    private void setCookieForToken(Map<String, String> tokens, HttpServletResponse response) {
-        tokens.forEach((index, token) -> {
-            Cookie newCookieToAdd = new Cookie(index, token);
-            newCookieToAdd.setHttpOnly(true);
-            newCookieToAdd.setMaxAge(cookieTime);
-            newCookieToAdd.setPath("/");
-            response.addCookie(newCookieToAdd);
-        });
-    }
 
     /*
-    *  SecurityConfig의 securityFilterChain() .logout() 설정
-    *  으로 인해서 이 컨트롤러 메서드는 실행되지 않음
-    * */
+     *  로그아웃
+     *  POST
+     *  SecurityConfig의 securityFilterChain() .logout() 설정으로 인해서 이 컨트롤러 메서드는 실행되지 않음
+     * */
     @PostMapping("/logout")
     public String deleteUser(HttpServletResponse response) {
         Optional<Authentication> authState = userService.logOut();
@@ -179,170 +125,11 @@ public class UserController {
 
 
     /*
-    * 마이페이지 - 프로필 조회
-    * */
-    @GetMapping("/settings/profile")
-    public String getProfileForm(@CookieValue("access_token") Cookie token, Model model) {
-        try {
-            String accessToken = token.getValue();
-            log.info("@CookieValue Cookie's access_token: {}", accessToken);
-
-            ProfileResponse profileData = userService.getProfileData(accessToken);
-
-            model.addAttribute(profileData);
-            return "/user/mypage/profile";
-
-        } catch (Exception e) {
-            log.error(e.getMessage());
-            return "redirect:/user/login";
-        }
-    }
-
-    /*
-    * 마이페이지 - 프로필 수정
-    * nickname
-    * img: S3 url
-    *
-    * */
-    @PutMapping("/settings/profile")
-    public ResponseEntity<?> updateProfile(@CookieValue("access_token") Cookie token,
-                                           @Valid ProfileRequest profileRequest, BindingResult bindingResult) {
-
-        try {
-            checkValidationAndReturnException(bindingResult);
-
-            String accessToken = token.getValue();
-
-            userService.updateProfileData(accessToken, profileRequest);
-            return new ResponseEntity<>(HttpStatus.OK);
-
-        } catch (Exception e) {
-            log.error("프로필 수정 실패: {}", e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /*
-    * 마이페이지 - 프로필 사진 삭제
-    * */
-    @DeleteMapping("/settings/profile/deleteimg")
-    public ResponseEntity<HttpStatus> deleteProfileImg(@CookieValue("access_token") Cookie token) {
-        log.info("deleteProfileImg 컨트롤러 실행");
-
-        try {
-            String accessToken = token.getValue();
-
-            userService.deleteProfileImg(accessToken);
-
-            return new ResponseEntity<>(HttpStatus.OK);
-
-        } catch (Exception e) {
-            log.error("이미지 삭제 실패: {}", e.getMessage());
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
-    }
-
-
-    /*
-    * 비밀번호 수정
-    * */
-    @GetMapping("/settings/profile/pwd")
-    public String getUpdatePasswordForm() {
-        return "/user/mypage/changePwdForm";
-    }
-    @PutMapping("/settings/profile/pwd")
-    public ResponseEntity<?> updatePassword(@CookieValue("access_token") Cookie token,
-                                 @Valid PasswordRequest passwordRequest, BindingResult bindingResult) {
-
-        try {
-            checkValidationAndReturnException(bindingResult);
-
-            userService.updatePassword(token.getValue(), passwordRequest);
-            return new ResponseEntity<>(HttpStatus.OK);
-
-        } catch (Exception e) {
-            log.error("비밀번호 수정 실패: {}", e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /*
-    * 이메일 수정
-    * */
-    @PutMapping("/email/change")
-    public ResponseEntity<?> updateEmail(@CookieValue("access_token") Cookie token,
-                                         @RequestBody @Valid MailChangeRequest mailChangeRequest, BindingResult bindingResult) {
-        try {
-            checkValidationAndReturnException(bindingResult);
-
-            userService.updateEmail(token.getValue(), mailChangeRequest);
-            return new ResponseEntity<>(HttpStatus.OK);
-
-        } catch (Exception e) {
-            log.error("이메일 변경 실패: {}", e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-
-    /*
-    * 아이디 찾기 이메일 전송
-    * */
-    @PostMapping("/email/find-id")
-    public ResponseEntity<?> findId(@RequestBody @Valid MailSendRequest mailSendRequest, BindingResult bindingResult) {
-        try {
-            checkValidationAndReturnException(bindingResult);
-
-            userService.sendMailWithId(mailSendRequest);
-            return new ResponseEntity<>(HttpStatus.OK);
-
-        } catch (Exception e) {
-            log.error("아이디 찾기 실패: {}", e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-    /*
-    * 비밀번호 찾기 이메일 전송
-    * */
-    @PostMapping("/email/find-pwd")
-    public ResponseEntity<?> findPwd(@RequestBody @Valid MailSendForPwdRequest mailSendForPwdRequest, BindingResult bindingResult) {
-        try {
-            checkValidationAndReturnException(bindingResult);
-
-            userService.checkUserAndSendVerifyCode(mailSendForPwdRequest);
-            return new ResponseEntity<>(HttpStatus.OK);
-
-        } catch (Exception e) {
-            log.error("아이디 찾기 실패: {}", e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-
-    /*
-    * 랜덤 비밀번호 설정: 이메일 전송
-    * */
-    @PostMapping("/email/change-pwd")
-    public ResponseEntity<?> changePwd(@RequestBody @Valid MailSendForPwdRequest mailSendForPwdRequest, BindingResult bindingResult) {
-        try {
-            checkValidationAndReturnException(bindingResult);
-
-            userService.changePwdRandomlyAndSendMail(mailSendForPwdRequest);
-            return new ResponseEntity<>(HttpStatus.OK);
-
-        } catch (Exception e) {
-            log.error("비밀번호 재설정 실패: {}", e.getMessage());
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-        }
-    }
-
-
-    /*
-    * 회원 탈퇴
-    *
-    * */
-    @PostMapping("/settings/profile/quit")
+     * 회원 탈퇴
+     * POST
+     *
+     * */
+    @PostMapping("/quit")
     public ResponseEntity<?> deleteUser(@CookieValue("access_token") Cookie token) {
         try {
             userService.withdrawUser(token.getValue());
@@ -354,59 +141,29 @@ public class UserController {
         }
     }
 
+
+
+
+
+    // 토큰을 쿠키에 저장
+    private void setCookieForToken(Map<String, String> tokens, HttpServletResponse response) {
+        tokens.forEach((index, token) -> {
+            Cookie newCookieToAdd = new Cookie(index, token);
+            newCookieToAdd.setHttpOnly(true);
+            newCookieToAdd.setMaxAge(cookieTime);
+            newCookieToAdd.setPath("/");
+            response.addCookie(newCookieToAdd);
+        });
+    }
+
+    private static String bindErrorPage(BindingResult bindingResult, Model model) {
+        model.addAttribute("errorMsg", bindingResult.getFieldError().getDefaultMessage());
+        return "/error/errorTemp";
+    }
+
     /*
-     * 마이페이지 - 작성글 목록: 커뮤니티
-     *
+     * DTO Validation 에러 체크 후 에러 발생시에러 메세지 세팅한 Exception throw
      * */
-    @GetMapping("/settings/profile/writings")
-    public String getMyWritingsPage(@CookieValue("access_token") Cookie token, Model model,
-                                    @RequestParam(name = "page", required = false, defaultValue = "0") Integer nowPage) {
-        try {
-            Pageable pageable = PageRequest.of(nowPage, 10, Sort.by("createdAt").descending());
-
-            Page<CommunityListResponseDto> data = userService.getMyCommunityWritings(token.getValue(), pageable);
-
-            model.addAttribute("nowPage", nowPage);
-            model.addAttribute("endPage", data.getTotalPages());
-            model.addAttribute("posts", data);
-            return "/user/mypage/myWritings";
-
-        } catch (Exception e) {
-            log.error("커뮤니티 작성글 불러오기 실패: {}", e.getMessage());
-            model.addAttribute("errorMsg", e.getMessage());
-            return "/error/errorTemp";
-        }
-    }
-
-    /*
-    * 마이페이지 - 북마크 목록: 여행/커뮤니티
-    *
-    * */
-    @GetMapping("/settings/profile/bookmarks")
-    public String getMyBookmarkPage(@CookieValue("access_token") Cookie token, Model model,
-                                    @RequestParam(name = "type", required = false, defaultValue = "trip") String type,
-                                    @RequestParam(name = "page", required = false, defaultValue = "0") Integer nowPage) {
-        try {
-            Pageable pageable = PageRequest.of(nowPage, 10, Sort.by("createdAt").descending());
-            Page<?> data = userService.getMyBookmarks(token.getValue(), type, pageable);
-
-            model.addAttribute("type", type);
-            model.addAttribute("nowPage", nowPage);
-            model.addAttribute("endPage", data.getTotalPages());
-            model.addAttribute("posts", data);
-            return "/user/mypage/myBookmarks";
-
-        } catch (Exception e) {
-            log.error("북마크 불러오기 실패: {}", e.getMessage());
-            model.addAttribute("errorMsg", e.getMessage());
-            return "/error/errorTemp";
-        }
-    }
-
-
-    /*
-    * DTO Validation 에러 체크 후 에러 발생시에러 메세지 세팅한 Exception throw
-    * */
     private static void checkValidationAndReturnException(BindingResult bindingResult) throws Exception {
         if (bindingResult.hasErrors()) {
             log.error("실패: {}", bindingResult.getFieldError().getDefaultMessage());
@@ -414,9 +171,4 @@ public class UserController {
         }
     }
 
-
-    private static String bindErrorPage(BindingResult bindingResult, Model model) {
-        model.addAttribute("errorMsg", bindingResult.getFieldError().getDefaultMessage());
-        return "/error/errorTemp";
-    }
 }
