+++ src/test/java/com/donguri/jejudorang/domain/user/api/UserControllerTest.java
@@ -26,7 +26,8 @@ class UserControllerTest {
     @Autowired EntityManager em;
     @Autowired Validator validator;
 
-    @Autowired UserController userController;
+    @Autowired
+    User2Controller userController;
 
     @Autowired UserService userService;
     @Autowired RedisTemplate<String, String> redisTemplate;
