+++ src/test/java/com/donguri/jejudorang/domain/user/repository/UserRepositoryTest.java
@@ -1,6 +1,7 @@
 package com.donguri.jejudorang.domain.user.repository;
 
 import com.donguri.jejudorang.domain.user.entity.*;
+import com.donguri.jejudorang.domain.user.entity.auth.Authentication;
 import com.donguri.jejudorang.domain.user.entity.auth.Password;
 import com.donguri.jejudorang.domain.user.repository.auth.PasswordRepository;
 import jakarta.persistence.EntityManager;
